+++ src/webpack/plugins/open-host-url/host-config.ts
@@ -1,0 +1,17 @@
+/**
+ * Configuration that is passed to the SKY UX Host server.
+ */
+export interface SkyuxHostConfig {
+
+  /**
+   * An array of JavaScript file names to inject into Host's index.html file.
+   */
+  scripts: {
+    name: string;
+  }[];
+
+  /**
+   * The URL of the localhost server, serving the asset files.
+   */
+  localUrl: string;
+}
