+++ src/webpack/plugins/open-host-url/open-host-url-config.ts
@@ -1,0 +1,12 @@
+export interface SkyuxOpenHostURLPluginConfig {
+
+  /**
+   * The URL of the SKY UX Host server.
+   */
+  hostUrl: string;
+
+  /**
+   * The URL of the localhost server, serving the asset files.
+   */
+  localUrl: string;
+}
